==> Code/01-data-documentation.qmd <==
---
title: "Data-documentation"
format: 
  html:
    embed-resources: true
html-math-method: 
  method: katex
editor: source
execute:
  warning: false
---

## Description of the raw data

The data files that are analysed in this paper were manually created (00-prepare-raw-data) from the following raw data files which we have added to the repository for reference:

+ pool-community-biomass-data.csv
+ pool-community-abundance-data.csv
+ pool-sediment-data.csv

These files used for analysis are separated by taxa group (i.e. all taxa, active dispersers and passive dispersers) and by ecological scale (i.e. *gamma-scale* which is the scale of the whole inselberg and the *alpha-scale* which has the data for individual pools).

+ gamma-all-raw.csv (separator: ;)
+ gamma-active-raw.csv (separator: ;)
+ gamma-passive-raw.csv (separator: ;)

+ alpha-all-raw.csv (separator: ;)
+ alpha-active-raw.csv (separator: ;)
+ alpha-passive-raw.csv (separator: ;)

### Gamma data variable descriptions

The three gamma-scale datasets have the same variables which are described below:

#### Inselberg:

1. type (i.e. double, integer, character): Charachter
2. unit: NA
3. description: Three letter code specifying the inselberg from which the data were gathered. 

#### Gamma:

1. type (i.e. double, integer, character): Double
2. unit: NA
3. description: Gamma diversity for each inselberg as estimated using the Chao-1 estimator based on incidence frequency.

#### Biomass(median):

1. type (i.e. double, integer, character): Double
2. unit: mg/L
3. description: Community-level invertebrate biomass per litre.

#### PC1:

1. type (i.e. double, integer, character): Double 
2. unit: NA
3. description: PC1 axis scores of the principal component analysis (PCA) of 19 different bioclimatic variables for each inselberg. Bioclimatic variables were extracted from the WorldClim database (Fick & Hijmans, 2017) at a resolution of 0.04167°. 

#### PC2:

1. type (i.e. double, integer, character): Double
2. unit: NA
3. description: PC2 axis scores of the principal component analysis (PCA) of 19 different bioclimatic variables for each inselberg. Bioclimatic variables were extracted from the WorldClim database (Fick & Hijmans, 2017) at a resolution of 0.04167°. 

#### Depth(median):

1. type (i.e. double, integer, character): Double
2. unit: cm
3. description: Estimated depth for each rock pool. 

#### Slope:

1. type (i.e. double, integer, character): Double
2. unit: NA
3. description: The relationship between standardised biomass and standardised alpha diversity on each inselberg, measured as the slope coefficient from the following linear model: ‘logbiomass_std ~ logdepth_std + Inselberg + logalpha_std:Inselberg’. 

#### SSE:

1. type (i.e. double, integer, character): Double
2. unit: NA
3. description: The standard error corresponding each slope coefficient. 

#### Prominence:

1. type (i.e. double, integer, character): Double
2. unit: m
3. description: Height of an inselberg above the surrounding landscape. 

#### GSE:

1. type (i.e. double, integer, character): Double
2. unit: NA
3. description: Standard error corresponding to the gamma Chao diversity. 


### Alpha data variable descriptions

The three alpha-scale datasets have the same variables which are described below:

#### Inselberg:

1. type (i.e. double, integer, character): Charachter
2. unit: NA
3. description: Three letter code specifying the inselberg from which the data were gathered. 

#### Pool:

1. type (i.e. double, integer, character): Charachter
2. unit: NA
3. description: Identity of the specific rock pool sampled. 

#### Alpha:

1. type (i.e. double, integer, character): Double
2. unit: NA
3. description: Alpha diversity for each rock pool as estimated using the Chao-1 estimator based on abundance.

#### Biomass:

1. type (i.e. double, integer, character): Double
2. unit: mg/L
3. description: Community-level invertebrate biomass per litre.

#### Depth:

1. type (i.e. double, integer, character): Double
2. unit: cm
3. description: Estimated depth for each rock pool.

#### Gamma:

1. type (i.e. double, integer, character): Double
2. unit: NA
3. description: Gamma diversity for each inselberg as estimated using the Chao-1 estimator based on incidence frequency. 

#### PC1:

1. type (i.e. double, integer, character): Double 
2. unit: NA
3. description: PC1 axis scores of the principal component analysis (PCA) of 19 different bioclimatic variables for each inselberg. 

#### PC2:

1. type (i.e. double, integer, character): Double
2. unit: NA
3. description: PC2 axis scores of the principal component analysis (PCA) of 19 different bioclimatic variables for each inselberg.

#### ASE:

1. type (i.e. double, integer, character): Double
2. unit: NA
3. description: Standard error corresponding to the alpha Chao diversity. 

#### GSE:

1. type (i.e. double, integer, character): Double
2. unit: NA
3. description: Standard error corresponding to the gamma Chao diversity.
