==> Code/07-combine-h1-plots.qmd <==
---
title: "Combine hypothesis 1 plots"
format: 
  html:
    embed-resources: true
html-math-method: 
  method: katex
editor: source
execute:
  warning: false
---

This notebook loads the plots relevant for hypothesis 1 and combines them into a single plot:

```{r}
# load relevant libraries
library(ggpubr)
library(ggplot2)

# get a list of .rds plots
files <- list.files(here::here("Figures/"))

# main text: fig. 5

# get relevant file names
fig5 <- files[ grepl(pattern = "_5[a-z]", files) ]
print(fig5)

# load the files for fig. 5
fig5 <- lapply(fig5, function(x) readRDS(here::here(paste0("Figures/", x) ) ))

# remove background from legend
fig5 <- lapply(fig5, function(x) {
  x + theme(legend.key = element_blank())
})

# combine using ggarrange
fig5 <- 
  ggarrange(plotlist = fig5, ncol = 3, nrow = 1, labels = c("(a)", "(b)", "(c)"),
            font.label = list(size = 11, face = "plain"),
            widths = c(1.1, 1, 1),
            common.legend = TRUE, legend = "bottom") 
fig5
```

```{r}
# export the plot
ggsave(here::here("Figures/fig_5.pdf"), fig5, units = "cm", width = 20,
       height = 9.5)
```
