==> Code/15-sediment-depth-comparison.qmd <==
---
title: "Sediment-depth comparison"
format: 
  html:
    embed-resources: true
html-math-method: 
  method: katex
editor: source
execute:
  warning: false
---

This notebook documents that analysis of the comparison between sediment depth and the depth of the pool.

```{r}
# load the relevant libraries
library(dplyr)
library(ggplot2)
library(readr)
library(wesanderson)

# load plotting theme
source(here::here("Code/helper-plotting-theme.R"))
```

We start by loading and cleaning the raw data:

```{r}
# load the sediment data
sed_dat <- readr::read_csv2(here::here("Data-raw/pool-sediment-data.csv"))
head(sed_dat)

# convert the lndepth and lnsed variables to numeric variables
sed_dat <- 
  sed_dat |>
  dplyr::mutate(Lndepth = as.numeric(Lndepth),
                LnSedPerVRatio = as.numeric(LnSedPerVRatio))

# rename the Location column
sed_dat <- dplyr::rename(sed_dat, Inselberg = Location)

# check the inselberg names
unique(sed_dat$Inselberg)
sed_dat <- 
  sed_dat |>
  dplyr::mutate(Inselberg = ifelse(Inselberg == "SOZ", "SWA",
                                   ifelse(Inselberg == "NOZ", "NWA", Inselberg)))

# convert inselberg to a factor
sed_dat <- dplyr::mutate(sed_dat, Inselberg = factor(Inselberg))

# sort out the coloours because not all inselbergs are present here
inselbergs <- c("IVC","SPN","KAM","SWE","FRA","USA","KOR","SWA","NWA","MAL")
inselbergs <- sort(inselbergs)
cols <- wesanderson::wes_palette(name = "Darjeeling1", n = 10, type = "continuous")

# which inselbergs are present
cols <- cols[which(inselbergs %in% unique(sed_dat$Inselberg))]

# perform a correlation test on these data
sed_cor <- cor.test(sed_dat$Lndepth, sed_dat$LnSedPerVRatio, method="pearson") # r = -0.73; P<0.001
sed_est <- round(sed_cor$estimate, 2)
```

Next, we plot the figure for the supplementary material:

```{r}
# plot fig. s5.1
p1 <- 
  ggplot(data = sed_dat, mapping = aes(x = Lndepth, y = LnSedPerVRatio, colour = Inselberg)) +
  geom_point(shape = 1, stroke = 0.5, alpha = 0.8, size = 1.75) +
  labs(x = "Ln( depth ) (cm)", y = "Ln( sediment/volume ) (% cover/m³)")+
  scale_colour_manual(values = cols) +
  guides(size = "none", 
         colour = guide_legend(override.aes = list(size = 1.5, 
                                                   alpha = 1, shape = 1, stroke = 1),
                                                   ncol = 1)) +
  annotate("text", x = Inf, y = Inf, hjust = 1.1, vjust = 1.5,
           label = paste0("r = ", sed_est, " *")) +
  theme_meta() +
  theme(legend.position = "right")+
  theme(legend.key = element_blank())

plot(p1)
```

Finally, we export the plot:

```{r}
# export the figure
ggsave(here::here("Figures/fig_s5.1.pdf"), p1, 
       units = "cm", width = 12, height = 9.5)
```
